backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-docx==1.1.0
openai==1.3.7
pyannote.audio==3.1.1
torch==2.1.1
torchaudio==2.1.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
jinja2==3.1.2
openai-whisper==20231117
pytest==7.4.3
pytest-asyncio==0.21.1
# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
# Cloud storage
cloudinary==1.36.0
# Additional for production
gunicorn==21.2.0
requests==2.31.0
